[tests] Fix examples tests with shared util
